Add vanilla CSS support to NavigationMenu (#533)
* Add / update hero demos

* Update examples
