+++ components/demos/NavigationMenu/stitches/index.jsx
@@ -1,0 +1,357 @@
+import React from 'react';
+import * as NavigationMenu from '@radix-ui/react-navigation-menu';
+import { styled, keyframes } from '@stitches/react';
+import { CaretDownIcon } from '@radix-ui/react-icons';
+import { violet, mauve, indigo, purple, blackA } from '@radix-ui/colors';
+
+const NavigationMenuDemo = () => {
+  return (
+    <NavigationMenuRoot>
+      <NavigationMenuList>
+        <NavigationMenu.Item>
+          <NavigationMenuTrigger>
+            Learn <CaretDown aria-hidden />
+          </NavigationMenuTrigger>
+          <NavigationMenuContent>
+            <List>
+              <li style={{ gridRow: 'span 3' }}>
+                <NavigationMenu.Link asChild>
+                  <Callout href="/">
+                    <svg aria-hidden width="38" height="38" viewBox="0 0 25 25" fill="white">
+                      <path d="M12 25C7.58173 25 4 21.4183 4 17C4 12.5817 7.58173 9 12 9V25Z"></path>
+                      <path d="M12 0H4V8H12V0Z"></path>
+                      <path d="M17 8C19.2091 8 21 6.20914 21 4C21 1.79086 19.2091 0 17 0C14.7909 0 13 1.79086 13 4C13 6.20914 14.7909 8 17 8Z"></path>
+                    </svg>
+                    <CalloutHeading>Radix Primitives</CalloutHeading>
+                    <CalloutText>Unstyled, accessible components for React.</CalloutText>
+                  </Callout>
+                </NavigationMenu.Link>
+              </li>
+
+              <ListItem href="https://stitches.dev/" title="Stitches">
+                CSS-in-JS with best-in-class developer experience.
+              </ListItem>
+              <ListItem href="/colors" title="Colors">
+                Beautiful, thought-out palettes with auto dark mode.
+              </ListItem>
+              <ListItem href="https://icons.radix-ui.com/" title="Icons">
+                A crisp set of 15x15 icons, balanced and consistent.
+              </ListItem>
+            </List>
+          </NavigationMenuContent>
+        </NavigationMenu.Item>
+
+        <NavigationMenu.Item>
+          <NavigationMenuTrigger>
+            Overview <CaretDown aria-hidden />
+          </NavigationMenuTrigger>
+          <NavigationMenuContent>
+            <List layout="two">
+              <ListItem title="Introduction" href="/docs/primitives/overview/introduction">
+                Build high-quality, accessible design systems and web apps.
+              </ListItem>
+              <ListItem title="Getting started" href="/docs/primitives/overview/getting-started">
+                A quick tutorial to get you up and running with Radix Primitives.
+              </ListItem>
+              <ListItem title="Styling" href="/docs/primitives/overview/styling">
+                Unstyled and compatible with any styling solution.
+              </ListItem>
+              <ListItem title="Animation" href="/docs/primitives/overview/animation">
+                Use CSS keyframes or any animation library of your choice.
+              </ListItem>
+              <ListItem title="Accessibility" href="/docs/primitives/overview/accessibility">
+                Tested in a range of browsers and assistive technologies.
+              </ListItem>
+              <ListItem title="Releases" href="/docs/primitives/overview/releases">
+                Radix Primitives releases and their changelogs.
+              </ListItem>
+            </List>
+          </NavigationMenuContent>
+        </NavigationMenu.Item>
+
+        <NavigationMenu.Item>
+          <NavigationMenuLink href="https://github.com/radix-ui">Github</NavigationMenuLink>
+        </NavigationMenu.Item>
+
+        <NavigationMenuIndicator>
+          <Arrow />
+        </NavigationMenuIndicator>
+      </NavigationMenuList>
+
+      <ViewportPosition>
+        <NavigationMenuViewport />
+      </ViewportPosition>
+    </NavigationMenuRoot>
+  );
+};
+
+const enterFromRight = keyframes({
+  from: { transform: 'translateX(200px)', opacity: 0 },
+  to: { transform: 'translateX(0)', opacity: 1 },
+});
+
+const enterFromLeft = keyframes({
+  from: { transform: 'translateX(-200px)', opacity: 0 },
+  to: { transform: 'translateX(0)', opacity: 1 },
+});
+
+const exitToRight = keyframes({
+  from: { transform: 'translateX(0)', opacity: 1 },
+  to: { transform: 'translateX(200px)', opacity: 0 },
+});
+
+const exitToLeft = keyframes({
+  from: { transform: 'translateX(0)', opacity: 1 },
+  to: { transform: 'translateX(-200px)', opacity: 0 },
+});
+
+const scaleIn = keyframes({
+  from: { transform: 'rotateX(-30deg) scale(0.9)', opacity: 0 },
+  to: { transform: 'rotateX(0deg) scale(1)', opacity: 1 },
+});
+
+const scaleOut = keyframes({
+  from: { transform: 'rotateX(0deg) scale(1)', opacity: 1 },
+  to: { transform: 'rotateX(-10deg) scale(0.95)', opacity: 0 },
+});
+
+const fadeIn = keyframes({
+  from: { opacity: 0 },
+  to: { opacity: 1 },
+});
+
+const fadeOut = keyframes({
+  from: { opacity: 1 },
+  to: { opacity: 0 },
+});
+
+const NavigationMenuRoot = styled(NavigationMenu.Root, {
+  position: 'relative',
+  display: 'flex',
+  justifyContent: 'center',
+  width: '100vw',
+  zIndex: 1,
+});
+
+const NavigationMenuList = styled(NavigationMenu.List, {
+  display: 'flex',
+  justifyContent: 'center',
+  backgroundColor: 'white',
+  padding: 4,
+  borderRadius: 6,
+  listStyle: 'none',
+  boxShadow: `0 2px 10px ${blackA.blackA7}`,
+  margin: 0,
+});
+
+const itemStyles = {
+  padding: '8px 12px',
+  outline: 'none',
+  userSelect: 'none',
+  fontWeight: 500,
+  lineHeight: 1,
+  borderRadius: 4,
+  fontSize: 15,
+  color: violet.violet11,
+  '&:focus': { boxShadow: `0 0 0 2px ${violet.violet7}` },
+  '&:hover': { backgroundColor: violet.violet3 },
+};
+
+const NavigationMenuTrigger = styled(NavigationMenu.Trigger, {
+  all: 'unset',
+  ...itemStyles,
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  gap: 2,
+});
+
+const NavigationMenuLink = styled(NavigationMenu.Link, {
+  ...itemStyles,
+  display: 'block',
+  textDecoration: 'none',
+  fontSize: 15,
+  lineHeight: 1,
+});
+
+const NavigationMenuContent = styled(NavigationMenu.Content, {
+  position: 'absolute',
+  top: 0,
+  left: 0,
+  width: '100%',
+  '@media only screen and (min-width: 600px)': { width: 'auto' },
+  '@media (prefers-reduced-motion: no-preference)': {
+    animationDuration: '250ms',
+    animationTimingFunction: 'ease',
+    '&[data-motion="from-start"]': { animationName: enterFromLeft },
+    '&[data-motion="from-end"]': { animationName: enterFromRight },
+    '&[data-motion="to-start"]': { animationName: exitToLeft },
+    '&[data-motion="to-end"]': { animationName: exitToRight },
+  },
+});
+
+const NavigationMenuIndicator = styled(NavigationMenu.Indicator, {
+  display: 'flex',
+  alignItems: 'flex-end',
+  justifyContent: 'center',
+  height: 10,
+  top: '100%',
+  overflow: 'hidden',
+  zIndex: 1,
+
+  '@media (prefers-reduced-motion: no-preference)': {
+    transition: 'width, transform 250ms ease',
+    '&[data-state="visible"]': { animation: `${fadeIn} 200ms ease` },
+    '&[data-state="hidden"]': { animation: `${fadeOut} 200ms ease` },
+  },
+});
+
+const NavigationMenuViewport = styled(NavigationMenu.Viewport, {
+  position: 'relative',
+  transformOrigin: 'top center',
+  marginTop: 10,
+  width: '100%',
+  backgroundColor: 'white',
+  borderRadius: 6,
+  overflow: 'hidden',
+  boxShadow: 'hsl(206 22% 7% / 35%) 0px 10px 38px -10px, hsl(206 22% 7% / 20%) 0px 10px 20px -15px',
+  height: 'var(--radix-navigation-menu-viewport-height)',
+
+  '@media only screen and (min-width: 600px)': {
+    width: 'var(--radix-navigation-menu-viewport-width)',
+  },
+  '@media (prefers-reduced-motion: no-preference)': {
+    transition: 'width, height, 300ms ease',
+    '&[data-state="open"]': { animation: `${scaleIn} 200ms ease` },
+    '&[data-state="closed"]': { animation: `${scaleOut} 200ms ease` },
+  },
+});
+
+const List = styled('ul', {
+  display: 'grid',
+  padding: 22,
+  margin: 0,
+  columnGap: 10,
+  listStyle: 'none',
+  variants: {
+    layout: {
+      one: {
+        '@media only screen and (min-width: 600px)': {
+          width: 500,
+          gridTemplateColumns: '.75fr 1fr',
+        },
+      },
+      two: {
+        '@media only screen and (min-width: 600px)': {
+          width: 600,
+          gridAutoFlow: 'column',
+          gridTemplateRows: 'repeat(3, 1fr)',
+        },
+      },
+    },
+  },
+  defaultVariants: {
+    layout: 'one',
+  },
+});
+
+const ListItem = React.forwardRef(({ children, title, ...props }, forwardedRef) => (
+  <li>
+    <NavigationMenu.Link asChild>
+      <ListItemLink {...props} ref={forwardedRef}>
+        <ListItemHeading>{title}</ListItemHeading>
+        <ListItemText>{children}</ListItemText>
+      </ListItemLink>
+    </NavigationMenu.Link>
+  </li>
+));
+
+const ListItemLink = styled('a', {
+  display: 'block',
+  outline: 'none',
+  textDecoration: 'none',
+  userSelect: 'none',
+  padding: 12,
+  borderRadius: 6,
+  fontSize: 15,
+  lineHeight: 1,
+  '&:focus': { boxShadow: `0 0 0 2px ${violet.violet7}` },
+  '&:hover': { backgroundColor: mauve.mauve3 },
+});
+
+const ListItemHeading = styled('div', {
+  fontWeight: 500,
+  lineHeight: 1.2,
+  marginBottom: 5,
+  color: violet.violet12,
+});
+
+const ListItemText = styled('p', {
+  all: 'unset',
+  color: mauve.mauve11,
+  lineHeight: 1.4,
+  fontWeight: 'initial',
+});
+
+const Callout = styled('a', {
+  display: 'flex',
+  justifyContent: 'flex-end',
+  flexDirection: 'column',
+  width: '100%',
+  height: '100%',
+  background: `linear-gradient(135deg, ${purple.purple9} 0%, ${indigo.indigo9} 100%);`,
+  borderRadius: 6,
+  padding: 25,
+  textDecoration: 'none',
+  outline: 'none',
+  userSelect: 'none',
+  '&:focus': { boxShadow: `0 0 0 2px ${violet.violet7}` },
+});
+
+const CalloutHeading = styled('div', {
+  color: 'white',
+  fontSize: 18,
+  fontWeight: 500,
+  lineHeight: 1.2,
+  marginTop: 16,
+  marginBottom: 7,
+});
+
+const CalloutText = styled('p', {
+  all: 'unset',
+  color: mauve.mauve4,
+  fontSize: 14,
+  lineHeight: 1.3,
+});
+
+const ViewportPosition = styled('div', {
+  position: 'absolute',
+  display: 'flex',
+  justifyContent: 'center',
+  width: '100%',
+  top: '100%',
+  left: 0,
+  perspective: '2000px',
+});
+
+const CaretDown = styled(CaretDownIcon, {
+  position: 'relative',
+  color: violet.violet10,
+  top: 1,
+  '[data-state=open] &': { transform: 'rotate(-180deg)' },
+  '@media (prefers-reduced-motion: no-preference)': {
+    transition: 'transform 250ms ease',
+  },
+});
+
+const Arrow = styled('div', {
+  position: 'relative',
+  top: '70%',
+  backgroundColor: 'white',
+  width: 10,
+  height: 10,
+  transform: 'rotate(45deg)',
+  borderTopLeftRadius: 2,
+});
+
+export default NavigationMenuDemo;
