+++ components/demos/index.jsx
@@ -10,7 +10,7 @@ export { default as DialogDemo } from './Dialog/stitches';
 export { default as DropdownMenuDemo } from './DropdownMenu/stitches';
 export { default as HoverCardDemo } from './HoverCard/stitches';
 export { default as LabelDemo } from './Label/stitches';
-export { default as NavigationMenuDemo } from './NavigationMenu';
+export { default as NavigationMenuDemo } from './NavigationMenu/stitches';
 export { default as PopoverDemo } from './Popover/stitches';
 export { default as ProgressDemo } from './Progress/stitches';
 export { default as RadioGroupDemo } from './RadioGroup/stitches';
