+++ data/primitives/components/navigation-menu/1.1.1.mdx
@@ -13,9 +13,7 @@ aria: https://www.w3.org/TR/wai-aria/#navigation
   <NavigationMenuDemo />
 </HeroContainer>
 
-```jsx hero template=NavigationMenu
-
-```
+<HeroCodeBlock folder="NavigationMenu" />
 
 <Highlights
   features={[
@@ -587,18 +585,9 @@ Tab focus will be maintained automatically.
 You can use the optional `Indicator` part to highlight the currently active `Trigger`, this is useful when you want to provide
 an animated visual cue such as an arrow or highlight to accompany the `Viewport`.
 
-```jsx line=4-11,25
-import { styled } from '@stitches/react';
+```jsx line=16
+// index.jsx
 import * as NavigationMenu from '@radix-ui/react-navigation-menu';
-
-const StyledIndicator = styled(NavigationMenu.Link, {
-  backgroundColor: 'grey',
-
-  '&[data-orientation="horizontal"]': { height: 3 },
-  '@media (prefers-reduced-motion: no-preference)': {
-    transition: 'width, transform, 250ms ease',
-  },
-});
 
 export default () => (
   <NavigationMenu.Root>
@@ -612,12 +601,27 @@ export default () => (
         <NavigationMenu.Content>Item two content</NavigationMenu.Content>
       </NavigationMenu.Item>
 
-      <StyledIndicator />
+      <NavigationMenu.Indicator __className__="NavigationMenuIndicator" />
     </NavigationMenu.List>
 
     <NavigationMenu.Viewport />
   </NavigationMenu.Root>
 );
+```
+
+```css
+/* styles.css */
+.NavigationMenuIndicator {
+  background-color: grey;
+}
+.NavigationMenuIndicator[data-orientation='horizontal'] {
+  height: 3px;
+}
+@media (prefers-reduced-motion: no-preference) {
+  .NavigationMenuIndicator {
+    transition: width, transform, 250ms ease;
+  }
+}
 ```
 
 ### With submenus
@@ -663,16 +667,11 @@ sure to assign and set a `defaultValue`.
 If you need to use the `Link` component provided by your routing package then we recommend composing with `NavigationMenu.Link` via a custom component.
 This will ensure accessibility and consistent keyboard control is maintained. Here's an example using Stitches and Next.js:
 
-```jsx line=11-20,26,29
-import { styled } from '@stitches/react';
+```jsx line=6-15,21,24
+// index.jsx
 import { useRouter } from 'next/router'
 import Link as NextLink from 'next/link';
 import * as NavigationMenu from '@radix-ui/react-navigation-menu';
-
-const StyledNavigationMenuLink = styled(NavigationMenu.Link, {
-  textDecoration: 'none',
-  '&[data-active]': { textDecoration: 'underline' },
-});
 
 const Link = ({ href, ...props }) => {
   const router = useRouter();
@@ -680,7 +679,7 @@ const Link = ({ href, ...props }) => {
 
   return (
     <NextLink href={href} passHref>
-      <StyledNavigationMenuLink active={isActive} {...props} />
+      <NavigationMenu.Link className="NavigationMenuLink" active={isActive} {...props} />
     </NextLink>
   );
 };
@@ -699,6 +698,16 @@ export default () => (
 );
 ```
 
+```css
+/* styles.css */
+.NavigationMenuLink {
+  text-decoration: none;
+}
+.NavigationMenuLink[data-active] {
+  text-decoration: 'underline';
+}
+```
+
 ### Advanced animation
 
 We expose `--radix-navigation-menu-viewport-[width|height]` and `data-motion['from-start'|'to-start'|'from-end'|'to-end']` attributes
@@ -707,71 +716,112 @@ to allow you to animate `Viewport` size and `Content` position based on the ente
 Combining these with `position: absolute;` allows you to create smooth overlapping animation effects when moving between items.
 Here's an example using Stitches:
 
-```jsx line=30-35,45,54,58,62
-import { styled, keyframes } from '@stitches/react';
+```jsx line=9-11,15-17,21
+// index.jsx
 import * as NavigationMenu from '@radix-ui/react-navigation-menu';
-
-const enterFromLeft = css.keyframes({
-  from: { transform: 'translateX(-100%)', opacity: 0 },
-  to: { transform: 'translateX(0)', opacity: 1 },
-});
-
-const enterFromRight = css.keyframes({
-  from: { transform: 'translateX(100%)', opacity: 0 },
-  to: { transform: 'translateX(0)', opacity: 1 },
-});
-
-const exitToLeft = css.keyframes({
-  from: { transform: 'translateX(0)', opacity: 1 },
-  to: { transform: 'translateX(-100%)', opacity: 0 },
-});
-
-const exitToRight = css.keyframes({
-  from: { transform: 'translateX(0)', opacity: 1 },
-  to: { transform: 'translateX(100%)', opacity: 0 },
-});
-
-const StyledContent = styled(NavigationMenu.Root, {
-  position: 'absolute',
-  top: 0,
-  left: 0,
-
-  '@media (prefers-reduced-motion: no-preference)': {
-    animationDuration: '250ms',
-    animationTimingFunction: 'ease',
-    '&[data-motion="from-start"]': { animationName: enterFromLeft },
-    '&[data-motion="from-end"]': { animationName: enterFromRight },
-    '&[data-motion="to-start"]': { animationName: exitToLeft },
-    '&[data-motion="to-end"]': { animationName: exitToRight },
-  },
-});
-
-const StyledViewport = styled(NavigationMenu.Viewport, {
-  position: 'relative',
-  width: 'var(--radix-hover-menu-viewport-width)',
-  height: 'var(--radix-hover-menu-viewport-height)',
-
-  '@media (prefers-reduced-motion: no-preference)': {
-    transition: 'width, height, 250ms ease',
-  },
-});
 
 export default () => (
   <NavigationMenu.Root>
     <NavigationMenu.List>
       <NavigationMenu.Item>
         <NavigationMenu.Trigger>Item one</NavigationMenu.Trigger>
-        <StyledContent>Item one content</StyledContent>
+        <NavigationMenu.Content __className__="NavigationMenuContent">
+          Item one content
+        </NavigationMenu.Content>
       </NavigationMenu.Item>
       <NavigationMenu.Item>
         <NavigationMenu.Trigger>Item two</NavigationMenu.Trigger>
-        <StyledContent>Item two content</StyledContent>
+        <NavigationMenu.Content __className__="NavigationMenuContent">
+          Item two content
+        </NavigationMenu.Content>
       </NavigationMenu.Item>
     </NavigationMenu.List>
 
-    <StyledViewport />
+    <NavigationMenu.Viewport __className__="NavigationMenuViewport" />
   </NavigationMenu.Root>
 );
+```
+
+```css line=12-23,28,29
+/* styles.css */
+.NavigationMenuContent {
+  position: absolute;
+  top: 0;
+  left: 0;
+}
+@media (prefers-reduced-motion: no-preference) {
+  .NavigationMenuContent {
+    animation-duration: 250ms;
+    animation-timing-function: ease;
+  }
+  .NavigationMenuContent[__data-motion__='from-start'] {
+    animation-name: enterFromLeft;
+  }
+  .NavigationMenuContent[__data-motion__='from-end'] {
+    animation-name: enterFromRight;
+  }
+  .NavigationMenuContent[__data-motion__='to-start'] {
+    animation-name: exitToLeft;
+  }
+  .NavigationMenuContent[__data-motion__='to-end'] {
+    animation-name: exitToRight;
+  }
+}
+
+.NavigationMenuViewport {
+  position: relative;
+  width: var(__--radix-navigation-menu-viewport-width__);
+  height: var(__--radix-navigation-menu-viewport-height__);
+}
+@media (prefers-reduced-motion: no-preference) {
+  .NavigationMenuViewport {
+    transition: width, height, 250ms ease;
+  }
+}
+
+@keyframes enterFromRight {
+  from {
+    opacity: 0;
+    transform: translateX(200px);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+
+@keyframes enterFromLeft {
+  from {
+    opacity: 0;
+    transform: translateX(-200px);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+
+@keyframes exitToRight {
+  from {
+    opacity: 1;
+    transform: translateX(0);
+  }
+  to {
+    opacity: 0;
+    transform: translateX(200px);
+  }
+}
+
+@keyframes exitToLeft {
+  from {
+    opacity: 1;
+    transform: translateX(0);
+  }
+  to {
+    opacity: 0;
+    transform: translateX(-200px);
+  }
+}
 ```
 
 ## Accessibility
